+++ src/api/get_2_tweets_id_liking_users.rs
@@ -1,5 +1,5 @@
 use crate::fields::{tweet_fields::TweetFields, user_fields::UserFields};
-use crate::responses::{errors::Errors, includes::Includes, users::Users};
+use crate::responses::{errors::Errors, includes::Includes, meta::Meta, users::Users};
 use crate::{api::execute_twitter, error::Error, rate_limit::RateLimit};
 use itertools::Itertools;
 use reqwest::RequestBuilder;
@@ -138,6 +138,7 @@ pub struct Response {
     pub data: Option<Vec<Users>>,
     pub errors: Option<Vec<Errors>>,
     pub includes: Option<Includes>,
+    pub meta: Option<Meta>,
     #[serde(flatten)]
     pub extra: std::collections::HashMap<String, serde_json::Value>,
 }
@@ -157,6 +158,11 @@ impl Response {
                 .unwrap_or(true)
             && self
                 .includes
+                .as_ref()
+                .map(|it| it.is_empty_extra())
+                .unwrap_or(true)
+            && self
+                .meta
                 .as_ref()
                 .map(|it| it.is_empty_extra())
                 .unwrap_or(true);