+++ src/api/get_2_users_id_liked_tweets.rs
@@ -2,7 +2,7 @@ use crate::fields::{
     media_fields::MediaFields, place_fields::PlaceFields, poll_fields::PollFields,
     tweet_fields::TweetFields, user_fields::UserFields,
 };
-use crate::responses::{errors::Errors, includes::Includes, tweets::Tweets};
+use crate::responses::{errors::Errors, includes::Includes, meta::Meta, tweets::Tweets};
 use crate::{api::execute_twitter, error::Error, rate_limit::RateLimit};
 use itertools::Itertools;
 use reqwest::RequestBuilder;
@@ -198,6 +198,7 @@ pub struct Response {
     pub data: Option<Vec<Tweets>>,
     pub errors: Option<Vec<Errors>>,
     pub includes: Option<Includes>,
+    pub meta: Option<Meta>,
     #[serde(flatten)]
     pub extra: std::collections::HashMap<String, serde_json::Value>,
 }
@@ -217,6 +218,11 @@ impl Response {
                 .unwrap_or(true)
             && self
                 .includes
+                .as_ref()
+                .map(|it| it.is_empty_extra())
+                .unwrap_or(true)
+            && self
+                .meta
                 .as_ref()
                 .map(|it| it.is_empty_extra())
                 .unwrap_or(true);