+++ tests/get_2_users_by_username_username.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use twapi_v2::api::{execute_twitter, get_2_users_by_username_username};
 
-// BEARER_CODE=XXXXX USER_NAME=XXXXX cargo test test_get_2_users_by_username_username -- --nocapture --test-threads=1
+// BEARER_CODE=XXXXX cargo test test_get_2_users_by_username_username -- --nocapture --test-threads=1
 
 #[tokio::test]
 async fn test_get_2_users_by_username_username() -> Result<()> {
@@ -12,6 +12,7 @@ async fn test_get_2_users_by_username_username() -> Result<()> {
     let response = serde_json::from_value::<get_2_users_by_username_username::Response>(res)?;
     let data = response.data.as_ref().unwrap();
     assert_eq!(data.username, "aoyagikouhei");
+    assert_eq!(data.id, "19522946");
     assert_eq!(response.is_empty_extra(), true);
     Ok(())
 }